orchestration/runtime/mwaa/requirements.txt
apache-airflow-providers-openlineage==1.4.0
apache-airflow-providers-openlineage[common.sql]==1.4.0
openlineage-integration-common==1.7.0
openlineage-python==1.7.0
openlineage_sql==1.7.0